Add cryptography function for creating base32 address from public key
