+++ elements/lisk-cryptography/src/keys.ts
@@ -14,7 +14,11 @@
  */
 import { bufferToHex, hexToBuffer } from './buffer';
 // eslint-disable-next-line import/no-cycle
-import { getAddressFromPublicKey } from './convert';
+import {
+	getAddressFromPublicKey,
+	convertUInt5ToBase32,
+	convertUIntArray,
+} from './convert';
 import { hash } from './hash';
 // eslint-disable-next-line import/no-cycle
 import { getKeyPair, getPublicKey } from './nacl';
@@ -80,4 +84,52 @@ export const getAddressFromPrivateKey = (privateKey: string): string => {
 	const publicKey = bufferToHex(publicKeyBytes);
 
 	return getAddressFromPublicKey(publicKey);
+};
+
+const GENERATOR = [0x3b6a57b2, 0x26508e6d, 0x1ea119fa, 0x3d4233dd, 0x2a1462b3];
+
+const polymod = (uint5Array: number[]): number => {
+	let chk = 1;
+	for (const value of uint5Array) {
+		// eslint-disable-next-line no-bitwise
+		const top = chk >> 25;
+		// eslint-disable-next-line no-bitwise
+		chk = ((chk & 0x1ffffff) << 5) ^ value;
+		for (let i = 0; i < 5; i += 1) {
+			// eslint-disable-next-line no-bitwise
+			if ((top >> i) & 1) {
+				// eslint-disable-next-line no-bitwise
+				chk ^= GENERATOR[i];
+			}
+		}
+	}
+
+	return chk;
+};
+
+export const createChecksum = (uint5Array: number[]): number[] => {
+	const values = uint5Array.concat([0, 0, 0, 0, 0, 0]);
+	// eslint-disable-next-line no-bitwise
+	const mod = polymod(values) ^ 1;
+	const result = [];
+	for (let p = 0; p < 6; p += 1) {
+		// eslint-disable-next-line no-bitwise
+		result.push((mod >> (5 * (5 - p))) & 31);
+	}
+	return result;
+};
+
+export const getBinaryAddressFromPublicKey = (publicKey: string): Buffer => {
+	const publicKeyBuffer = Buffer.from(publicKey, 'hex');
+	return hash(publicKeyBuffer).slice(0, 20);
+};
+
+export const getBase32AddressFromPublicKey = (
+	publicKey: string,
+	prefix: string,
+): string => {
+	const binaryAddress = getBinaryAddressFromPublicKey(publicKey);
+	const uint5Address = convertUIntArray(Uint8Array.from(binaryAddress), 8, 5);
+	const uint5Checksum = createChecksum(uint5Address);
+	return `${prefix}${convertUInt5ToBase32(uint5Address.concat(uint5Checksum))}`;
 };
